refactor(rent car): create rental/useCases/RentCar dir and move related file there
