+++ src/modules/rental/routes/rental.routes.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { RentCarController } from '../useCases/RentCarController';
+import { RentCarController } from '../useCases/RentCar/RentCarController';
 
 const rentCarController = new RentCarController();
